requirements.txt
streamlit
pandas
plotly
google-genai
python-dotenv
pillow
pyTelegramBotAPI
oracledb
google-cloud-secret-manager
